--- lib/contourpy/util/__init__.py
@@ -1,8 +1,0 @@
-from .data import random_uniform
-from .mpl_renderer import MplRenderer, MplTestRenderer, MplDebugRenderer
-
-try:
-    # Bokeh may not be available.  Fail silently.
-    from .bokeh_renderer import BokehRenderer
-except:
-    pass
